LayoutTests/http/wpt/web-animations/interfaces/KeyframeEffect/composite-expected.txt


FAIL Default value assert_equals: The default value should be replace expected (string) "replace" but got (undefined) undefined
PASS Change composite value 
FAIL Unspecified keyframe composite value when setting effect composite anim.effect.getKeyframes is not a function. (In 'anim.effect.getKeyframes()', 'anim.effect.getKeyframes' is undefined)
FAIL Specified keyframe composite value when setting effect composite anim.effect.getKeyframes is not a function. (In 'anim.effect.getKeyframes()', 'anim.effect.getKeyframes' is undefined)
